Moved other auth modules into feature folder
