+++ config/sebk2.json
@@ -1,0 +1,31 @@
+{
+  "agoraUrlRoot": "https://agora.sebk2.bee.envs-terra.bio",
+  "alphaRegionalityGroup": "Alpha_Regionality_Users",
+  "b2cBillingPolicy": "b2c_1a_signup_signin_billing_dev",
+  "bardRoot": "https://terra-bard-dev.appspot.com",
+  "billingProfileManagerUrlRoot": "https://bpm.sebk2.bee.envs-terra.bio",
+  "bondUrlRoot": "https://bond.sebk2.bee.envs-terra.bio",
+  "calhounUrlRoot": "https://calhoun.dsde-dev.broadinstitute.org",
+  "catalogUrlRoot": "https://catalog.dsde-dev.broadinstitute.org",
+  "dataRepoUrlRoot": "https://data.sebk2.bee.envs-terra.bio",
+  "devUrlRoot": "https://bvdp-saturn-dev.appspot.com",
+  "dockstoreUrlRoot": "https://staging.dockstore.org",
+  "externalCredsUrlRoot": "https://externalcreds.dsde-dev.broadinstitute.org",
+  "firecloudBucketRoot": "https://storage.googleapis.com/firecloud-alerts-dev",
+  "firecloudUrlRoot": "https://firecloudui.sebk2.bee.envs-terra.bio",
+  "jobManagerUrlRoot": "https://job-manager.sebk2.bee.envs-terra.bio/jobs",
+  "leoUrlRoot": "https://leonardo.sebk2.bee.envs-terra.bio",
+  "marthaUrlRoot": "https://martha.sebk2.bee.envs-terra.bio",
+  "orchestrationUrlRoot": "https://firecloudorch.sebk2.bee.envs-terra.bio",
+  "rawlsUrlRoot": "https://rawls.sebk2.bee.envs-terra.bio",
+  "rexUrlRoot": "https://terra-rex-dev.appspot.com",
+  "samUrlRoot": "https://sam.sebk2.bee.envs-terra.bio",
+  "shibbolethUrlRoot": "https://broad-shibboleth-prod.appspot.com/dev",
+  "tCell": {
+    "apiKey": "AQEBBAEkx4iE2KxNyI7Wx08EwU1ycTM7E4FMSmaibbMUQxNU6uQvuAJt7fyABAtFYSYfgEE",
+    "appId": "saturnnonprod-O9lUP"
+  },
+  "terraDockerImageBucket": "terra-docker-image-documentation-dev",
+  "terraDockerVersionsFile": "terra-docker-versions-dev.json",
+  "workspaceManagerUrlRoot": "https://workspace.sebk2.bee.envs-terra.bio"
+}
