+++ src/appLoader.js
@@ -5,8 +5,8 @@ import _ from 'lodash/fp';
 import { createRoot } from 'react-dom/client';
 import { h } from 'react-hyperscript-helpers';
 import RModal from 'react-modal';
+import { initializeAuth } from 'src/auth/auth';
 import { initializeClientId } from 'src/auth/oidc-broker';
-import { initializeAuth } from 'src/libs/auth';
 import { isAxeEnabled } from 'src/libs/config';
 import { startPollingServiceAlerts } from 'src/libs/service-alerts-polling';
 import { initializeTCell } from 'src/libs/tcell';
