+++ src/auth/oauth-redirect-loader.js
@@ -1,6 +1,6 @@
 import { createRoot } from 'react-dom/client';
 import { h } from 'react-hyperscript-helpers';
-import RedirectFromOAuth from 'src/pages/RedirectFromOAuth';
+import RedirectFromOAuth from 'src/auth/RedirectFromOAuth';
 
 const rootElement = document.getElementById('root');
 const root = createRoot(rootElement);
