+++ src/components/TopBar.js
@@ -4,6 +4,7 @@ import { Fragment, useRef, useState } from 'react';
 import { UnmountClosed as RCollapse } from 'react-collapse';
 import { a, div, h, h1, img, span } from 'react-hyperscript-helpers';
 import { Transition } from 'react-transition-group';
+import { signIn, signOut } from 'src/auth/auth';
 import AlertsIndicator from 'src/components/Alerts';
 import { Clickable, IdContainer, LabeledCheckbox, Link, spinnerOverlay } from 'src/components/common';
 import { icon } from 'src/components/icons';
@@ -15,7 +16,6 @@ import fcIconWhite from 'src/images/brands/firecloud/FireCloud-icon-white.svg';
 import headerLeftHexes from 'src/images/header-left-hexes.svg';
 import headerRightHexes from 'src/images/header-right-hexes.svg';
 import { Ajax } from 'src/libs/ajax';
-import { signIn, signOut } from 'src/libs/auth';
 import { isBaseline, isBioDataCatalyst, isDatastage, isFirecloud, isTerra } from 'src/libs/brand-utils';
 import colors from 'src/libs/colors';
 import { getConfig } from 'src/libs/config';
