+++ src/pages/Main.ts
@@ -6,9 +6,9 @@ import { h } from 'react-hyperscript-helpers';
 import { ReactNotifications } from 'react-notifications-component';
 import { AuthProvider } from 'react-oidc-context';
 import { AuthenticatedCookieSetter } from 'src/analysis/runtime-common-components';
+import AuthContainer from 'src/auth/AuthContainer';
+import AuthStoreSetter from 'src/auth/AuthStoreSetter';
 import { getOidcConfig } from 'src/auth/oidc-broker';
-import AuthContainer from 'src/components/AuthContainer';
-import AuthStoreSetter from 'src/components/AuthStoreSetter';
 import ConfigOverridesWarning from 'src/components/ConfigOverridesWarning';
 import CookieRejectModal from 'src/components/CookieRejectModal';
 import CookieWarning from 'src/components/CookieWarning';
