+++ src/pages/billing/Project.js
@@ -3,6 +3,7 @@ import * as qs from 'qs';
 import { Fragment, useEffect, useMemo, useState } from 'react';
 import { div, h, p, span } from 'react-hyperscript-helpers';
 import { cloudProviders } from 'src/analysis/utils/runtime-utils';
+import * as Auth from 'src/auth/auth';
 import { ButtonPrimary, customSpinnerOverlay, IdContainer, Link, Select } from 'src/components/common';
 import { DeleteUserModal, EditUserModal, MemberCard, MemberCardHeaders, NewUserCard, NewUserModal } from 'src/components/group-common';
 import { icon } from 'src/components/icons';
@@ -14,7 +15,6 @@ import { MenuTrigger } from 'src/components/PopupTrigger';
 import { SimpleTabBar } from 'src/components/tabBars';
 import { ariaSort, HeaderRenderer } from 'src/components/table';
 import { Ajax } from 'src/libs/ajax';
-import * as Auth from 'src/libs/auth';
 import colors from 'src/libs/colors';
 import { reportErrorAndRethrow } from 'src/libs/error';
 import Events, { extractBillingDetails } from 'src/libs/events';
