+++ src/pages/workspaces/workspace/JobHistory.js
@@ -2,6 +2,7 @@ import _ from 'lodash/fp';
 import { Fragment, useImperativeHandle, useRef, useState } from 'react';
 import { div, h, span, table, tbody, td, tr } from 'react-hyperscript-helpers';
 import { AutoSizer } from 'react-virtualized';
+import { bucketBrowserUrl } from 'src/auth/auth';
 import * as breadcrumbs from 'src/components/breadcrumbs';
 import { Clickable, Link, spinnerOverlay } from 'src/components/common';
 import { icon } from 'src/components/icons';
@@ -13,7 +14,6 @@ import { MenuTrigger } from 'src/components/PopupTrigger';
 import { FlexTable, HeaderRenderer, TextCell, TooltipCell } from 'src/components/table';
 import TooltipTrigger from 'src/components/TooltipTrigger';
 import { Ajax } from 'src/libs/ajax';
-import { bucketBrowserUrl } from 'src/libs/auth';
 import colors from 'src/libs/colors';
 import { reportError } from 'src/libs/error';
 import * as Nav from 'src/libs/nav';
