+++ src/profile/notification-settings/NotificationSettings.js
@@ -1,12 +1,12 @@
 import _ from 'lodash/fp';
 import { useEffect, useState } from 'react';
 import { div, h, h2 } from 'react-hyperscript-helpers';
+import { refreshTerraProfile } from 'src/auth/auth';
 import { Checkbox, spinnerOverlay } from 'src/components/common';
 import { InfoBox } from 'src/components/InfoBox';
 import { PageBox, PageBoxVariants } from 'src/components/PageBox';
 import { useWorkspaces } from 'src/components/workspace-utils';
 import { Ajax } from 'src/libs/ajax';
-import { refreshTerraProfile } from 'src/libs/auth';
 import { withErrorReporting } from 'src/libs/error';
 import Events from 'src/libs/events';
 import { memoWithName } from 'src/libs/react-utils';
