+++ src/workspace-data/data-table/uri-viewer/UriViewer.js
@@ -3,6 +3,7 @@ import filesize from 'filesize';
 import _ from 'lodash/fp';
 import { Fragment, useState } from 'react';
 import { div, h, p, pre, span } from 'react-hyperscript-helpers';
+import { bucketBrowserUrl } from 'src/auth/auth';
 import { requesterPaysWrapper, withRequesterPaysHandler } from 'src/components/bucket-utils';
 import { ClipboardButton } from 'src/components/ClipboardButton';
 import Collapse from 'src/components/Collapse';
@@ -10,7 +11,6 @@ import { Link } from 'src/components/common';
 import { parseGsUri } from 'src/components/data/data-utils';
 import Modal from 'src/components/Modal';
 import { Ajax } from 'src/libs/ajax';
-import { bucketBrowserUrl } from 'src/libs/auth';
 import colors from 'src/libs/colors';
 import { isFeaturePreviewEnabled } from 'src/libs/feature-previews';
 import { useCancellation, useOnMount, withDisplayName } from 'src/libs/react-utils';
